+++ apps/aspiration/models/__init__.py
@@ -1,0 +1,3 @@
+from .aspirations import Aspiration, AspirationProgress, AspirationFile
+from .categories import  Category
+from .notifications import Notification